+++ instrumentation/sql/vendor_postgres.go
@@ -6,9 +6,16 @@ import (
 	nurl "net/url"
 	"sort"
 	"strings"
+	"unicode"
 )
 
 type postgresExtension struct{}
+
+// scanner implements a tokenizer for libpq-style option strings.
+type scanner struct {
+	s []rune
+	i int
+}
 
 func init() {
 	vendorExtensions = append(vendorExtensions, &postgresExtension{})
@@ -16,9 +23,7 @@ func init() {
 
 // Gets if the extension is compatible with the component name
 func (ext *postgresExtension) IsCompatible(componentName string) bool {
-	return componentName == "pq.Driver" ||
-		componentName == "stdlib.Driver" ||
-		componentName == "pgsqldriver.postgresDriver"
+	return componentName == "pq.Driver" || componentName == "stdlib.Driver" || componentName == "pgsqldriver.postgresDriver"
 }
 
 // Complete the missing driver data from the connection string
@@ -34,7 +39,7 @@ func (ext *postgresExtension) ProcessConnectionString(connectionString string, c
 	o := make(values)
 	o["host"] = "localhost"
 	o["port"] = "5432"
-	_ = parseOpts(dsn, o)
+	_ = ext.parseOpts(dsn, o)
 	o["password"] = "******"
 
 	if user, ok := o["user"]; ok {
@@ -122,4 +127,109 @@ func (ext *postgresExtension) parseUrl(url string) (string, error) {
 
 	sort.Strings(kvs) // Makes testing easier (not a performance concern)
 	return strings.Join(kvs, " "), nil
+}
+
+// parseOpts parses the options from name and adds them to the values.
+//
+// The parsing code is based on conninfo_parse from libpq's fe-connect.c
+func (ext *postgresExtension) parseOpts(name string, o values) error {
+	s := ext.newScanner(name)
+
+	for {
+		var (
+			keyRunes, valRunes []rune
+			r                  rune
+			ok                 bool
+		)
+
+		if r, ok = s.SkipSpaces(); !ok {
+			break
+		}
+
+		// Scan the key
+		for !unicode.IsSpace(r) && r != '=' {
+			keyRunes = append(keyRunes, r)
+			if r, ok = s.Next(); !ok {
+				break
+			}
+		}
+
+		// Skip any whitespace if we're not at the = yet
+		if r != '=' {
+			r, ok = s.SkipSpaces()
+		}
+
+		// The current character should be =
+		if r != '=' || !ok {
+			return fmt.Errorf(`missing "=" after %q in connection info string"`, string(keyRunes))
+		}
+
+		// Skip any whitespace after the =
+		if r, ok = s.SkipSpaces(); !ok {
+			// If we reach the end here, the last value is just an empty string as per libpq.
+			o[string(keyRunes)] = ""
+			break
+		}
+
+		if r != '\'' {
+			for !unicode.IsSpace(r) {
+				if r == '\\' {
+					if r, ok = s.Next(); !ok {
+						return fmt.Errorf(`missing character after backslash`)
+					}
+				}
+				valRunes = append(valRunes, r)
+
+				if r, ok = s.Next(); !ok {
+					break
+				}
+			}
+		} else {
+		quote:
+			for {
+				if r, ok = s.Next(); !ok {
+					return fmt.Errorf(`unterminated quoted string literal in connection string`)
+				}
+				switch r {
+				case '\'':
+					break quote
+				case '\\':
+					r, _ = s.Next()
+					fallthrough
+				default:
+					valRunes = append(valRunes, r)
+				}
+			}
+		}
+
+		o[string(keyRunes)] = string(valRunes)
+	}
+
+	return nil
+}
+
+// newScanner returns a new scanner initialized with the option string s.
+func (ext *postgresExtension) newScanner(s string) *scanner {
+	return &scanner{[]rune(s), 0}
+}
+
+// Next returns the next rune.
+// It returns 0, false if the end of the text has been reached.
+func (s *scanner) Next() (rune, bool) {
+	if s.i >= len(s.s) {
+		return 0, false
+	}
+	r := s.s[s.i]
+	s.i++
+	return r, true
+}
+
+// SkipSpaces returns the next non-whitespace rune.
+// It returns 0, false if the end of the text has been reached.
+func (s *scanner) SkipSpaces() (rune, bool) {
+	r, ok := s.Next()
+	for unicode.IsSpace(r) && ok {
+		r, ok = s.Next()
+	}
+	return r, ok
 }
